+++ NEWS.md
@@ -1,6 +1,11 @@
 # document 2.0.0.9000
 
-man() now works with RStudio.
+* man() now works with RStudio.
+* add and export alter\_description(). 
+  This serves as a replacement for the internal
+  add\_depenedencies\_to\_description(), and straightens the internal
+  clean\_descprition() which used [write|read]Lines() instead of
+  [write|read].dcf().
 
 # document 2.0.0